pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "asr_qe"
version = "0.1.0"
description = "ASR Quality Estimation System"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "numpy<2.0",
    "librosa",
    "soundfile",
    "datasets==2.20.0",
    "xgboost",
    "scipy",
    "joblib",
    "pandas",
    "hydra-core",
    "omegaconf",
    "transformers",
    "jiwer",
    "tqdm",
    # "nemo_toolkit[asr]", -- Moved to optional to speed up build
    # "cuda-python>=12.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "prometheus-client>=0.23.1",
    "fastapi",
    "uvicorn",
    "python-multipart",
]

[project.optional-dependencies]
research = [
    "nemo_toolkit[asr]",
    "cuda-python>=12.3",
]
dev = [
    "pytest",
    "pytest-mock",
    "ruff",
    "mypy",
    "black",
    "apache-airflow==2.9.1",
    "apache-airflow-providers-apache-spark",
]
spark = [
    "pyspark==3.5.0",
]
api = [
    "torch",
    "nemo_toolkit[asr]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
